Use posts page for search, if set
WordPress allows setting the "posts" page to be a page other than the
front page, something we use to host the blog at /blog.

Add a switch in the search form template which detects if this setting
is configured, and if so changes the search action to point to the posts
page rather than the home page.